# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'My API Documentation'
  description: 'API documentation for my Laravel application'
  contact:
    email: admin@example.com
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'API Server'
paths:
...
  /api/checkout:
    post:
      tags:
        - Orders
      summary: 'Checkout — converts all cart items into an order then clears the cart'
      operationId: checkout
      requestBody:
        required: false
        content:
          application/json:
            schema:
              properties:
                address:
                  description: 'Optional — shipping address, defaults to empty string if omitted'
                  type: string
                  example: '123 Main St, Phnom Penh'
                  nullable: true
              type: object
      responses:
        201:
          description: 'Order placed successfully — cart is cleared after this call'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  order:
                    properties:
                      id:
                        type: integer
                        example: 1
                      user_id:
                        type: integer
                        example: 1
                      total_amount:
                        type: number
                        format: float
                        example: 89.97
                      status:
                        description: 'Always "pending" on creation'
                        type: string
                        example: pending
                      address:
                        type: string
                        example: '123 Main St, Phnom Penh'
                      created_at:
                        type: string
                        format: date-time
                        example: '2024-01-01T00:00:00.000000Z'
                      items:
                        type: array
                        items:
                          properties:
                            id:
                              type: integer
                              example: 1
                            order_id:
                              type: integer
                              example: 1
                            product_id:
                              type: integer
                              example: 3
                            quantity:
                              type: integer
                              example: 2
                            price:
                              description: 'Price snapshot at the time of checkout'
                              type: number
                              format: float
                              example: 29.99
                            product:
                              properties:
                                id:
                                  type: integer
                                  example: 3
                                name:
                                  type: string
                                  example: 'Rose Perfume'
                                price:
                                  type: number
                                  format: float
                                  example: 29.99
                                image:
                                  type: string
                                  example: products/abc.jpg
                                  nullable: true
                              type: object
                          type: object
                    type: object
                type: object
        401:
          description: 'Unauthenticated — Bearer token missing or invalid'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
                type: object
        422:
          description: 'Cart is empty — nothing to checkout'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Cart is empty'
                type: object
        500:
          description: 'Checkout failed — transaction rolled back, cart is preserved'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Checkout failed: SQLSTATE[...] ...'
                type: object
      security:
        -
          sanctum: []
  /api/orders:
    get:
      tags:
        - Orders
      summary: 'Get all orders for the authenticated user (paginated, 10 per page)'
      operationId: getOrders
      responses:
        200:
          description: 'Orders retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Products are gotten successfully'
                  data:
                    properties:
                      current_page:
                        type: integer
                        example: 1
                      per_page:
                        type: integer
                        example: 10
                      total:
                        type: integer
                        example: 25
                      data:
                        type: array
                        items:
                          properties:
                            id:
                              type: integer
                              example: 1
                            user_id:
                              type: integer
                              example: 1
                            total_amount:
                              type: number
                              format: float
                              example: 89.97
                            status:
                              type: string
                              example: pending
                            address:
                              type: string
                              example: '123 Main St, Phnom Penh'
                            created_at:
                              type: string
                              format: date-time
                              example: '2024-01-01T00:00:00.000000Z'
                            items:
                              type: array
                              items:
                                properties:
                                  id:
                                    type: integer
                                    example: 1
                                  product_id:
                                    type: integer
                                    example: 3
                                  quantity:
                                    type: integer
                                    example: 2
                                  price:
                                    type: number
                                    format: float
                                    example: 29.99
                                  product:
                                    properties:
                                      id: { type: integer, example: 3 }
                                      name: { type: string, example: 'Rose Perfume' }
                                      price: { type: number, format: float, example: 29.99 }
                                      image: { type: string, example: products/abc.jpg, nullable: true }
                                    type: object
                                type: object
                          type: object
                    type: object
                type: object
        401:
          description: 'Unauthenticated — Bearer token missing or invalid'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
                type: object
      security:
        -
          sanctum: []
  '/api/orders/{id}':
    get:
      tags:
        - Orders
      summary: 'Get a single order with all its items and products'
      operationId: getOrder
      parameters:
        -
          name: id
          in: path
          description: 'Order ID — must belong to the authenticated user'
          required: true
          schema:
            type: integer
            example: 1
            minimum: 1
      responses:
        200:
          description: 'Order retrieved successfully'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Order is gotten successfully'
                  data:
                    properties:
                      id:
                        type: integer
                        example: 1
                      user_id:
                        type: integer
                        example: 1
                      total_amount:
                        type: number
                        format: float
                        example: 89.97
                      status:
                        type: string
                        example: pending
                      address:
                        type: string
                        example: '123 Main St, Phnom Penh'
                      created_at:
                        type: string
                        format: date-time
                        example: '2024-01-01T00:00:00.000000Z'
                      updated_at:
                        type: string
                        format: date-time
                        example: '2024-01-01T00:00:00.000000Z'
                      items:
                        type: array
                        items:
                          properties:
                            id:
                              type: integer
                              example: 1
                            order_id:
                              type: integer
                              example: 1
                            product_id:
                              type: integer
                              example: 3
                            quantity:
                              type: integer
                              example: 2
                            price:
                              description: 'Price snapshot at the time of checkout — may differ from current product price'
                              type: number
                              format: float
                              example: 29.99
                            product:
                              properties:
                                id:
                                  type: integer
                                  example: 3
                                name:
                                  type: string
                                  example: 'Rose Perfume'
                                price:
                                  type: number
                                  format: float
                                  example: 29.99
                                image:
                                  type: string
                                  example: products/abc.jpg
                                  nullable: true
                              type: object
                          type: object
                    type: object
                type: object
        401:
          description: 'Unauthenticated — Bearer token missing or invalid'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
                type: object
        404:
          description: 'Order not found or does not belong to the authenticated user'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Not found'
                type: object
      security:
        -
          sanctum: []
...
components:
...
  securitySchemes:
    sanctum:
      type: http
      description: 'Enter token in format: Bearer <token>'
      bearerFormat: JWT
      scheme: bearer
...